Prevent Evergreen from trying to get patron dumps
This is something that Millennium supports but Evergreen
does not. Trying to do it in evergreen causes errors and 
failure during login so our workaround is to not bother doing
it when authenticating with Evergreen. 

This is intended to be a temporary workaround which should
be overridden by NYPL's official Evergreen changes